+++ flashinfer/cute_dsl/utils.py
@@ -46,13 +46,13 @@ def cutlass_to_torch_dtype(cutlass_dtype):
     torch_dtype = getattr(torch, cutlass_dtype.__name__.lower(), None)
 
     torch_type_map = {
-        cutlass.dtypes.TFloat32: torch.float32,
-        cutlass.dtypes.Float32: torch.float32,
-        cutlass.dtypes.Float16: torch.float16,
-        cutlass.dtypes.BFloat16: torch.bfloat16,
-        cutlass.dtypes.Float8E5M2: torch.float8_e5m2,
-        cutlass.dtypes.Float8E4M3FN: torch.float8_e4m3fn,
-        cutlass.dtypes.Float8E4M3B11FNUZ: torch.float8_e4m3fnuz,
+        cutlass.TFloat32: torch.float32,
+        cutlass.Float32: torch.float32,
+        cutlass.Float16: torch.float16,
+        cutlass.BFloat16: torch.bfloat16,
+        cutlass.Float8E5M2: torch.float8_e5m2,
+        cutlass.Float8E4M3FN: torch.float8_e4m3fn,
+        cutlass.Float8E4M3B11FNUZ: torch.float8_e4m3fnuz,
     }
     if torch_dtype is None:
         torch_dtype = torch_type_map.get(cutlass_dtype)
